Improved/fixed collision to handle different player sizes, changed player x and y to be floats to allow smooth movement at low speeds.

+++ TDAHelper/library.hpp
@@ -46,7 +46,7 @@ namespace td {
 
     class Shapes {
     public:
-        static sf::RectangleShape rect(int x, int y, int width, int height, sf::Color color = sf::Color::White);
+        static sf::RectangleShape rect(float x, float y, int width, int height, sf::Color color = sf::Color::White);
         static sf::VertexArray line(int x1, int y1, int x2, int y2, sf::Color color = sf::Color::White);
     };
